adk-mcp-agentforce/requirements.txt
google-adk[a2a]
requests
python-dotenv
mcp
httpx-sse
httpx